requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration and logging
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0  # Optional: for experiment tracking

# Evaluation and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0

# Text processing
nltk>=3.8.0
regex>=2023.8.8

# Utilities
python-dotenv>=1.0.0